+++ cpr/auth.cpp
@@ -1,4 +1,4 @@
-#include "auth.h"
+#include "cpr/auth.h"
 
 namespace cpr {
